+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - endpoint support for /datastore/{datastoreId}/set_policy <POST>
     - endpoint support for /datastore/{datastoreId}/share <POST>
     - endpoint support for /datastore/{datastoreId}/standard_hosts <GET>
+    - endpoint support for /datastore/{datastoreId}/unshare <POST>
     - endpoint support for /external_stores <GET>
     - endpoint support for /external_stores <POST>
     - endpoint support for /external_stores/update_credentials <POST>